Solving the runtime problem with Olivier Gingras
Some of the modifications I made with Olivier when we were figuring out how to solve the runtime problem when we call sympyfy on the Strontium compound. It turns out I needed to establish a cut-off on the number of cells involved in the hopping, that is, not including all hoppings since some hopping amplitudes are but a fraction of nearest neighbor hoppings.