Established energy cutoff to w90 model
I made sure to establish an energy cutoff of 0.05 eV to the Wannier90 model, which is then converted to a TRIQS TB lattice object allowing us to pass it to the sympyfy function. There are some complications with the output, however. It seems like the analytical expression is not fully simplified which is evidenced by the fact that there are still lattice element symbols in the analytical expression. My best guess is that there is a problem when the dot product operation occurs.