Attempting to check whether Hamiltonian is Hermitian
Unsuccessful attempts to check whether the resulting Hamiltonian is Hermitian. This time I tried extracting the underlying values from the SymPy elements, filling in a normal NumPy array, and checking if that Hamiltonian is Hermitian. This was to be done mostly using the np.flatten method but it did not work out eventually.